+++ webots_ros2_driver/launch/webots_launch.py
@@ -24,12 +24,6 @@ import launch
 from webots_ros2_driver.webots_launcher import WebotsLauncher
 
 
-'''
-
-xacro ur.urdf.xacro > ur5e.urdf name:=ur5e joint_limit_params:=/home/benjamin/ros2_ws/src/webots_ros2/webots_ros2_universal_robot/resource/ur_description/config/ur5e/joint_limits.yaml kinematics_params:=/home/benjamin/ros2_ws/src/webots_ros2/webots_ros2_universal_robot/resource/ur_description/config/ur5e/default_kinematics.yaml physical_params:=/home/benjamin/ros2_ws/src/webots_ros2/webots_ros2_universal_robot/resource/ur_description/config/ur5e/physical_parameters.yaml visual_params:=/home/benjamin/ros2_ws/src/webots_ros2/webots_ros2_universal_robot/resource/ur_description/config/ur5e/visual_parameters.yaml
-
-'''
-
 def generate_launch_description():
     world = LaunchConfiguration('world')
 
@@ -37,7 +31,6 @@ def generate_launch_description():
         world=world,
     )
 
-    '''
     supervisor_spawner = Node(
         package='webots_ros2_driver',
         executable='supervisor_spawner.py',
@@ -45,13 +38,13 @@ def generate_launch_description():
         additional_env={'WEBOTS_ROBOT_NAME': 'Spawner'},
         respawn=True,
     )
-    '''
 
     return LaunchDescription([DeclareLaunchArgument(
             'world',
             description=f'Give a world file path as argument'
         ),
         webots,
+        supervisor_spawner,
         launch.actions.RegisterEventHandler(
             event_handler=launch.event_handlers.OnProcessExit(
                 target_action=webots,