+++ webots_ros2_driver/webots_ros2_driver/webots_launcher.py
@@ -18,74 +18,6 @@
 
 import os
 import sys
-from launch.actions import ExecuteProcess
-from launch.substitution import Substitution
-from launch.substitutions import TextSubstitution
-from webots_ros2_driver.utils import get_webots_home, handle_webots_installation
-
-
-class _ConditionalSubstitution(Substitution):
-    def __init__(self, *, condition, false_value='', true_value=''):
-        self.__condition = condition if isinstance(condition, Substitution) else TextSubstitution(text=str(condition))
-        self.__false_value = false_value if isinstance(false_value, Substitution) else TextSubstitution(text=false_value)
-        self.__true_value = true_value if isinstance(true_value, Substitution) else TextSubstitution(text=true_value)
-
-    def perform(self, context):
-        if context.perform_substitution(self.__condition).lower() in ['false', '0', '']:
-            return context.perform_substitution(self.__false_value)
-        return context.perform_substitution(self.__true_value)
-
-
-class WebotsLauncher(ExecuteProcess):
-    def __init__(self, output='screen', world=None, gui=True, mode='realtime', stream=False, **kwargs):
-        # Find Webots executable
-        webots_path = get_webots_home(show_warning=True)
-        if webots_path is None:
-            handle_webots_installation()
-            webots_path = get_webots_home()
-        if sys.platform == 'win32':
-            webots_path = os.path.join(webots_path, 'msys64', 'mingw64', 'bin')
-        webots_path = os.path.join(webots_path, 'webots')
-
-        mode = mode if isinstance(mode, Substitution) else TextSubstitution(text=mode)
-        world = world if isinstance(world, Substitution) else TextSubstitution(text=world)
-
-        no_rendering = _ConditionalSubstitution(condition=gui, false_value='--no-rendering')
-        stdout = _ConditionalSubstitution(condition=gui, false_value='--stdout')
-        stderr = _ConditionalSubstitution(condition=gui, false_value='--stderr')
-        no_sandbox = _ConditionalSubstitution(condition=gui, false_value='--no-sandbox')
-        if sys.platform == 'win32':
-            # Windows doesn't have the sandbox argument
-            no_sandbox = ''
-        minimize = _ConditionalSubstitution(condition=gui, false_value='--minimize')
-        stream_argument = _ConditionalSubstitution(condition=stream, true_value='--stream')
-        xvfb_run_prefix = []
-        if 'WEBOTS_OFFSCREEN' in os.environ:
-            xvfb_run_prefix.append('xvfb-run')
-            xvfb_run_prefix.append('--auto-servernum')
-            no_rendering = '--no-rendering'
-
-        # no_rendering, stdout, stderr, no_sandbox, minimize
-        super().__init__(
-            output=output,
-            cmd=xvfb_run_prefix + [
-                webots_path,
-                stream_argument,
-                no_rendering,
-                stdout,
-                stderr,
-                no_sandbox,
-                minimize,
-                world,
-                '--batch',
-                ['--mode=', mode],
-            ],
-            **kwargs
-        )
-
-'''
-import os
-import sys
 import shutil
 import tempfile
 
@@ -169,31 +101,28 @@ class WebotsLauncher(ExecuteProcess):
         if not world_path:
             sys.exit('World file not specified (has to be specified with world=absolute/path/to/my/world.wbt')
 
-        #''
+        '''
         if world_path[-len(URDF_world_suffix):] == URDF_world_suffix:
             world_copy = world_path
             world_path = world_path[:-len(URDF_world_suffix)] + '.wbt'
         else:
-        #''
+        '''
 
-        self.__world_copy = tempfile.NamedTemporaryFile(mode="w+", suffix=URDF_world_suffix)
+        self.__world_copy = tempfile.NamedTemporaryFile(mode="w+", suffix=URDF_world_suffix, delete=False)
         shutil.copy2(world_path, self.__world_copy.name)
         context.launch_configurations['world']=self.__world_copy.name
 
-        #''
         # Add supervisor Spawner
         indent = '  '
-        worldFile = open(self.__world_copy.name, 'a')
-        worldFile.write('Robot {\n')
-        worldFile.write(indent + 'name "Spawner"\n')
-        worldFile.write(indent + 'controller "<extern>"\n')
-        worldFile.write(indent + 'supervisor TRUE\n')
-        worldFile.write('}\n')
-        worldFile.close()
-        #''
+        world_file = open(self.__world_copy.name, 'a')
+        world_file.write('Robot {\n')
+        world_file.write(indent + 'name "Spawner"\n')
+        world_file.write(indent + 'controller "<extern>"\n')
+        world_file.write(indent + 'supervisor TRUE\n')
+        world_file.write('}\n')
+        world_file.close()
 
         return super().execute(context)
-
 
     def _shutdown_process(self, context, *, send_sigint):
         # Remove copy of the world and the corresponding ".wbproj" file
@@ -207,4 +136,3 @@ class WebotsLauncher(ExecuteProcess):
             if os.path.isfile(world_copy_secondary_file):
                 os.unlink(world_copy_secondary_file)
         return super()._shutdown_process(context, send_sigint=send_sigint)
-    '''